+++ core/Database/Concerns/Query/BuildsQuery.php
@@ -13,6 +13,7 @@ use Core\Database\Functions;
 use Core\Database\Having;
 use Core\Database\SelectCase;
 use Core\Database\Subquery;
+use Core\Database\Value;
 use Core\Util\Arr;
 
 trait BuildsQuery
@@ -194,10 +195,32 @@ trait BuildsQuery
         return $this;
     }
 
+    public function exists(): self
+    {
+        $this->action = Actions::SELECT_EXISTS;
+
+        $this->columns = [Operators::EXISTS->value];
+
+        return $this;
+    }
+
+    public function doesntExist(): self
+    {
+        $this->action = Actions::SELECT_EXISTS;
+
+        $this->columns = [Operators::NOT_EXISTS->value];
+
+        return $this;
+    }
+
+    /**
+     * @return array<int, mixed>
+     */
     public function toSql(): array
     {
         $sql = match ($this->action) {
             Actions::SELECT => $this->buildSelectQuery(),
+            Actions::SELECT_EXISTS => $this->buildSelectExistsQuery(),
             Actions::INSERT => $this->buildInsertSentence(),
             Actions::UPDATE => $this->buildUpdateSentence(),
             Actions::DELETE => $this->buildDeleteSentence(),
@@ -244,6 +267,23 @@ trait BuildsQuery
             $query[] = Arr::implodeDeeply($this->offset);
 
         }
+
+        return Arr::implodeDeeply($query);
+    }
+
+    protected function buildSelectExistsQuery(): string
+    {
+        $query = ['SELECT'];
+        $query[] = $this->columns[0];
+
+        $subquery[] = "SELECT 1 FROM {$this->table}";
+
+        if (! empty($this->clauses)) {
+            $subquery[] = 'WHERE';
+            $subquery[] = $this->prepareClauses($this->clauses);
+        }
+
+        $query[] = '(' . Arr::implodeDeeply($subquery) . ') AS ' . Value::from('exists');
 
         return Arr::implodeDeeply($query);
     }